refactor: move prepareHarborZone to harbor_zone.js and import in harbor.js
